Revert ":truck: move package 'service' into 'app'"
