+++ src/main/java/com/grepp/coffee/app/model/coffee/dto/CoffeeDto.java
@@ -1,16 +1,14 @@
-package com.grepp.coffee.app.model.dto;
+package com.grepp.coffee.app.model.coffee.dto;
 
-import lombok.AllArgsConstructor;
+import java.time.LocalDateTime;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
 @NoArgsConstructor
-@AllArgsConstructor
 public class CoffeeDto {
 
     private Integer coffeeId;
-    private String coffeeName;
     private Integer price;
     private Integer stock;
