--- src/main/java/com/grepp/coffee/app/model/entity/Coffee.java
@@ -1,15 +1,0 @@
-package com.grepp.coffee.app.model.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Coffee {
-    private int id;
-    private String name;
-    private int price;
-    private int stock;
-}
