+++ src/main/java/com/grepp/coffee/service/menu/MenuServiceImpl.java
@@ -1,0 +1,28 @@
+package com.grepp.coffee.service.menu;
+
+import com.grepp.coffee.app.model.coffee.dto.CoffeeDto;
+import org.springframework.stereotype.Service;
+
+@Service
+public class MenuServiceImpl implements MenuService {
+
+    @Override
+    public void addMenu(CoffeeDto coffeeDto) {
+
+    }
+
+    @Override
+    public CoffeeDto getCoffee() {
+        return null;
+    }
+
+    @Override
+    public void updateMenu(CoffeeDto coffeeDto) {
+
+    }
+
+    @Override
+    public void deleteMenu(CoffeeDto coffeeDto) {
+
+    }
+}
