+++ src/main/java/com/grepp/coffee/service/order/OrderService.java
@@ -1,6 +1,6 @@
-package com.grepp.coffee.app.service.order;
+package com.grepp.coffee.service.order;
 
-import com.grepp.coffee.app.model.dto.DetailedOrderDto;
+import com.grepp.coffee.app.model.detailedorder.dto.DetailedOrderDto;
 
 public interface OrderService {
     void processOrder(DetailedOrderDto detailedOrderDto);
